refactor: remove cli-parser logic into separate module
