+++ src/controller/controller.js
@@ -1,10 +1,7 @@
 import { Transform } from 'node:stream';
 
 import { printCurrentDirectory, upDirectory, setCurrentDirectory, listSortedDir } from '../directory/directory.js';
-
-const COMMAND_DELIMITER = ' ';
-const COMMAND_INDEX = 0;
-const FIRST_ARGS_INDEX = 1;
+import { parseUserCliCommand } from '../cli-parser/cli-parser.js';
 
 const COMMANDS = {
   EXIT: '.exit',
@@ -19,11 +16,8 @@ class Controller extends Transform {
   }
 
   async _transform(chunk, _, callback) {
-    let command;
     try {
-      const chunkNormalizedString = chunk.toString('utf8').trim();
-      const commandArgs = chunkNormalizedString.split(COMMAND_DELIMITER);
-      command = commandArgs[COMMAND_INDEX];
+      const { command, argument } = parseUserCliCommand(chunk);
 
       switch (command) {
         case COMMANDS.EXIT: 
@@ -31,20 +25,18 @@ class Controller extends Transform {
           break;
         case COMMANDS.UP_DIR:
           await upDirectory();
-          printCurrentDirectory();
           break;
         case COMMANDS.CHANGE_DIR: 
-          await setCurrentDirectory(commandArgs[FIRST_ARGS_INDEX]);
-          printCurrentDirectory();
+          await setCurrentDirectory(argument);
           break;
         case COMMANDS.LIST_DIR: 
           await listSortedDir();
-          printCurrentDirectory();
           break;
       }
     } catch(error) {
       console.log(error.message);
     } finally {
+      printCurrentDirectory();
       callback(null, chunk);
     }
   }
